Added verbose requip/removal events aimed for DD content mods usage.
These events provide an easy way to get notified whenever a device is
equipped/unequipped, when a key breaks or a lock gets jammed. In
contrast to the existing events, these work for all item types at once.
They also communicate the exact device that was manipulated and its
device keyword (in addition to the actor wearing it).