New Device Escape System
Added a new escape system. The new system supports device difficulty,
which can be set as a property in the item's script. It defaults to a
10% base chance to make a successful escape. The base chance (if greater
than zero) will get modifed with several bonuses if the player is
experienced in lockpicking, destruction or alteration magic, has
successfully escaped a number of restraints in the past, or has already
tried escaping this particularly restraint a lot. If the character fails
an attempt, she cannot try again for a while (configurable in MCM), and
there is a chance for a catastrophic failure (configurable in MCM) that
will prevent any further escape attempts. Last but not least there is a
difficulty modifier (you guessed right, it's configurable in MCM!)

The new system should be totally backwards compatible. If existing
custom items override the relevant library function, the old system (or
whatever the custom items replaced it with) will remain in use.  The old
system is still fully functional!

The new system is optional - I added a MCM toggle for it. It wouldn't be
me if I didn't!

If you want to make an item inescapable by any means other than
unlocking it with the proper key, just set its difficulty property to 0.
The modifiers will NOT apply in this case.