Added new zad_QuestItem keyword.
This new keyword should make handling quest items more convenient and
safer. Using it will make it easier to determine quest items as such, in
contrast to the zad_BlockGeneric keyword, which might or might not
indicate an actual quest item. Quest items tagged with this keyword can
not be removed via a generic call to libs.removeDevice(). They need
libs.removeQuestDevice() instead which among things take a non-framework
keyword present on the item as parameter. The removal will be refused if
no matching keyword can be found on the item. Bypassing their events is
also not possible (as this would often break things for these items).