Plug rework (phase two) - Lockable Plugs Framework
Added native support for lockable plugs to DDi.
Added two example plugs (locking anal and vaginal plugs) to serve as a
template.
Changed the messages in the original plug script to refer to plugs in
singular.
Changed the core equip script to move the RemovedWithSuccess condition
to a more useful place.