New device difficulty setting: LockAccessDifficulty
If set to greater than zero, the character cannot easily reach the locks
when locked in this restraint. The higher the number, the harder she
will find it to unlock herself, even when in possession of the key. A
value of 100 will make it impossible for her to reach the locks. She
will need help. Make sure that your mod actually provides a means to
escape such retraints!