Idle format overhaul
+ Purged the animation system with righteous fire and cleansed it of sin
and impurity
+ Also refactored all internal and device-end code to use
SendAnimationEvent instead of PlayIdle. Specific animations now must be
passed as strings (behaviour name)
+ Custom wrist restraints will need to be updated. the "StruggleIdle"
and "StruggleHobIdle" arrays are now string arrays
+ Hopefully fixed players occasionally getting offsets applied to them
even though it is and was physicially impossible unless someone's game
was borked beyond all recognition? But... I guess now it's
double-impossible?
+ Copied over a couple more ZAP animation files that are still in use
+ Hopefully erased all remaining traces of ZAP from DDi
+ Added horny, orgasm and edge animation support for elbowbinders, and
why-is-this-a-thing-yokes (placeholder animations, to be replaced in a
future update)
+ Frontcuffs use unbound animations because they kinda work as is
+ This update was sponsored by Dr Pepper