Added pony animations and new DestroyKey feature
Added AA sets for pony gear that can be triggered via a keyword (zad_DeviousPonyGear) on any device.

Reimplemented a function that can make any generic key be consumed on use and a MCM control for users to toggle it. This function is now quest-safe. Modders can allow it for their own keys by tagging them with the zad_NonUniqueKey keyword.